# Replace struct impl of Invariant by Arbitrary