--- a/library/kani/src/invariant.rs
+++ b/library/kani/src/invariant.rs
@@ -71,39 +71,6 @@ unsafe impl Invariant for char {
     }
 }
 
-unsafe impl<T: Invariant, const N: usize> Invariant for [T; N] {
-    fn is_valid(&self) -> bool {
-        self.iter().all(|e| e.is_valid())
-    }
-}
-
-unsafe impl<T> Invariant for Option<T>
-where
-    T: Invariant,
-{
-    #[inline(always)]
-    fn is_valid(&self) -> bool {
-        if let Some(v) = self { v.is_valid() } else { matches!(*self, None) }
-    }
-}
-
-unsafe impl<T, E> Invariant for Result<T, E>
-where
-    T: Invariant,
-    E: Invariant,
-{
-    #[inline(always)]
-    fn is_valid(&self) -> bool {
-        if let Ok(v) = self {
-            v.is_valid()
-        } else if let Err(e) = self {
-            e.is_valid()
-        } else {
-            false
-        }
-    }
-}
-
 macro_rules! nonzero_invariant {
     ( $type: ty ) => {
         unsafe impl Invariant for $type {