@celinval 2022-07-22
I forgot to mention that I did a few experiments with `check_both` test that I added. I changed the test to create an array with 100 elements instead.

For the arbitrary array, the performance impact with the changes in this PR was small. I tried a different implementation using MaybeUninit (similar to this example: https://doc.rust-lang.org/stable/std/mem/union.MaybeUninit.html#initializing-an-array-element-by-element) since this would be the safest, and there was a significant penalty on the verification time. Here is the data I collected:

| Before Changes  | After Changes | MaybeUninit |
| ------------- | ------------- | --------- |
| 10.0s | 10.7s | 12.7s |